将save方法添加至M层
