+++ thinkphp/application/index/model/Attraction.php
@@ -11,5 +11,25 @@ use think\Model;
  */
 
 class Attraction extends Model {
+    public function saveAttraction($title, $content, $name, $meal, $car, $guide, $image, $hotel, $article_id) {
+        $this->title = $title;
+        $this->description = $content;
+        $this->designation = $name;
+        $this->meal = $meal;
+        $this->car = $car;
+        $this->guide = $guide;
+        $this->image = $image;
+        $this->weight = Attraction::where('article_id', '=', $article_id)->count()+1;
+        $this->article_id = $article_id;
 
+        if(!is_null($hotel)){
+            $this->hotel_id = $hotel->id;
+        }
+
+        if(!$this->save()) {
+            return false;
+        } else {
+            return true;
+        }
+    }
 }
