--- thinkphp/application/index/model/AttractionModel.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace app\index\model;
-/**
- * Created by PhpStorm.
- * User: zhangxishuo
- * Date: 2017/8/30
- * Time: 20:49
- */
-
-class AttractionModel {
-    public function saveAttraction($title, $content, $name, $meal, $car, $guide, $image, $hotel, $article_id) {
-        $attraction = new Attraction();
-        $attraction->title = $title;
-        $attraction->description = $content;
-        $attraction->name = $name;
-        $attraction->meal = $meal;
-        $attraction->car = $car;
-        $attraction->guide = $guide;
-        $attraction->image = $image;
-        $attraction->weight = Attraction::where('article_id', '=', $article_id)->count()+1;
-        $attraction->article_id = $article_id;
-
-        if(!is_null($hotel)){
-            $attraction->hotel_id = $hotel->id;
-        }
-
-        if(!$attraction->save()) {
-            return false;
-        } else {
-            return true;
-        }
-    }
-}
